requirements.txt
streamlit
python-dotenv
PyPDF2==3.0.1
langchain
langchain_groq